requirements.txt
numpy==1.18.5
Pillow==6.1.0
opencv-python==4.1.1.26
tensorflow==2.3.1
Keras==2.3.0
RPi.GPIO==0.7.0
paho-mqtt==1.5.1